Update _includes/applescripts/status.applescript
I found that if a track contains double quotes it would fail to load the page and would produce javascript errors as the JSON array was then invalid.

I used the existing string_replace function in the status.applescript file to replace a doublequote with a singlequote in the titles of tracks. This resolved the issue.